requirements.txt
flask==2.3.3
langchain==0.1.1
langchain-core==0.1.10
langchain-community==0.0.13
python-dotenv==1.0.0 